requirements.txt
affine==2.3.0
astroid==2.3.3
attrs==19.3.0
awscli==1.16.156
botocore==1.12.146
certifi==2018.8.24
chardet==3.0.4
Click==7.0
click-plugins==1.1.1
cligj==0.5.0
colorama==0.3.9
cycler==0.10.0
docutils==0.14
greenlet==0.4.15
idna==2.7
isort==4.3.21
jmespath==0.9.4
kiwisolver==1.0.1
lazy-object-proxy==1.4.3
lxml==4.5.0
matplotlib==3.0.1
mccabe==0.6.1
msgpack==0.6.1
numpy==1.15.4
pandas==0.23.4
Pillow==7.1.1
pyasn1==0.4.5
pylint==2.4.4
-e git://github.com/ojii/pymaging.git@596a08fce5664e58d6e8c96847393fbe987783f2#egg=pymaging
-e git://github.com/ojii/pymaging-png.git@83d85c44e4b2342818e6c068065e031a9f81bb9f#egg=pymaging_png
pynvim==0.3.2
pyparsing==2.3.0
python-barcode==0.11.0
python-dateutil==2.7.5
pytz==2018.7
PyYAML==5.1
qrcode==6.1
rasterio==1.1.1
requests==2.20.0
rsa==3.4.2
s3transfer==0.2.0
scipy==1.1.0
seaborn==0.9.0
six==1.11.0
snuggs==1.4.7
termgraph==0.2.0
typed-ast==1.4.1
urllib3==1.24.2
wrapt==1.11.2
xlrd==1.2.0
